# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "bitgraph API",
    "version": "1.0.0",
    "description": "bit-graph task"
  },
  "servers": [
    {
      "url": "http://localhost:4000/"
    }
  ],
...
  "paths": {
...
    "/tasks/{id}": {
      "get": {
        "summary": "Get A Task By Id",
        "tags" : ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "شناسه تمرین"
          }
        ],
        "responses": {
          "200": {
            "description": "اطلاعات تمرین",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "404": { "description": "تمرین پیدا نشد" },
          "401": { "description": "نیاز به احراز هویت" },
          "500": { "description": "خطای سرور" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    },
      "components": {
        "securitySchemes": {
          "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        },
        "schemas": {
...
          "Task": {
            "type": "object",
            "properties": {
              "_id": { "type": "string", "example": "64ac4d77..." },
              "description": { "type": "string", "example": "تمرین شماره 1" },
              "file": { "type": "string", "example": "/tasks/file123.pdf" },
              "courseId": { "type": "string", "example": "64aaee1f..." },
              "userId": { "type": "string", "example": "64abbc33..." },
              "status": {
                "type": "string",
                "enum": ["بررسی شده", "بررسی نشده"],
                "default": "بررسی نشده"
              },
              "score": {
                "type": "number",
                "default": 0,
                "minimum": 0,
                "maximum": 10
              },
              "feedback": {
                "type": "string",
                "nullable": true,
                "example": "64addf11..."
              },
              "createdAt": { "type": "string", "format": "date-time" },
              "updatedAt": { "type": "string", "format": "date-time" }
            }
          },
...
        }
      }
    }